Move `FeedLoaderPresentationAdapter` to its own file
